Restauration des fichiers perdus

+++ Xmazon/Xmazon/model/Product.cs
@@ -5,14 +5,14 @@ namespace Xmazon
 {
 	public class Product
 	{
-		public string uid { get; set; }
-		public string name { get; set; }
-		public float price { get; set; }
-		public bool available { get; set; }
+		public string uid {get;set;}
+		public string name {get;set;}
+		public float price {get;set;}
+		public bool available {get;set;}
 
 		public Product(){}
 
-		public Product (string uid, string name, float price, bool available)
+		public Product(string uid, string name, float price, bool available)
 		{
 			this.uid = uid;
 			this.name = name;
@@ -20,11 +20,11 @@ namespace Xmazon
 			this.available = available;
 		}
 
-		public static Product Deserialize (JsonValue value)
+		public static Product Deserialize(JsonValue value)
 		{
 			Product product = new Product(value ["uid"],
 										  value ["name"],
-								(decimal) value["price"],
+								(float) value["price"],
 								   (bool) value["available"]);
 			return product;
